+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### 🚜 Refactor
+
+- Drop configurable proxy manager
+
 ## [0.9.0] - 2024-11-23
 
 ### 🚀 Features